+++ src/Tools.h
@@ -22,7 +22,7 @@ int dimension(string filename);
 int count_file_lines(string filename);
 
 /* h and g functions for LSH and hypercube calculations */
-int h_func(const vector<unsigned long> &p, int i);
-unsigned int g_func(const vector<unsigned long> &p, int i);
+int h_func(const vector<double> &p, int i);
+unsigned int g_func(const vector<double> &p, int i);
 
 #endif
